requirements.txt
annotated-types==0.7.0
anyio==4.9.0
Brotli==1.1.0
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cssselect2==0.8.0
et_xmlfile==2.0.0
fastapi==0.115.12
fonttools==4.58.1
h11==0.16.0
idna==3.10
Jinja2==3.1.6
lxml==5.4.0
Markdown==3.8
MarkupSafe==3.0.2
numpy==2.2.6
openpyxl==3.1.5
pandas==2.2.3
pdfkit==1.0.0
pillow==11.2.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydyf==0.11.0
PyMuPDF==1.26.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tinycss2==1.4.0
tinyhtml5==2.0.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uvicorn==0.34.3
weasyprint==65.1
webencodings==0.5.1
XlsxWriter==3.2.3
zopfli==0.2.3.post1
